requirements.txt
rich
prompt-toolkit
PyYAML